Move error tracking to hook
